refactor: split CLI smoke tests by feature
